+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* [THORN-335] Update to use patched pymodbus and add configurable pymodbus logging.
 * [THORN-382] Add a delay before the next poll after a device fails to respond.
 * [LOW-2008] Switch from Poetry to UV. Note this requires a .dockerignore file to ensure we don't pull in cached files resulting in the
 incorrect path to the Python interpreter (this caused the simulation pod not to start).